login_grant_type_user

+++ src/main/java/com/badminton/store/controller/AddressController.java
@@ -23,45 +23,30 @@ import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-
 import javax.validation.Valid;
 import java.util.List;
-import java.util.stream.Collectors;
-
-
 
 
 @RestController
 @RequestMapping("/v1/address")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class AddressController extends ABasicController{
-
     @Autowired
     private AddressRepository addressRepository;
-
     @Autowired
     private AddressMapper addressMapper;
-
     @Autowired
     private UserRepository userRepository;
-
     @Autowired
     private NationRepository nationRepository;
 
     @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
-    @PreAuthorize("hasRole('A-C')")
+    @PreAuthorize("hasRole('A_C')")
     public ApiMessageDto<String> create (@Valid @RequestBody CreateAddressForm form, BindingResult bindingResult){
         ApiMessageDto<String> apiMessageDto = new ApiMessageDto<>();
-        if (bindingResult.hasErrors()) {
-            // Duyệt qua tất cả các lỗi và nối thành một chuỗi
-            String errorMessage = bindingResult.getFieldErrors()
-                    .stream()
-                    .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                    .collect(Collectors.joining(", "));
-
-            throw new BadRequestException(errorMessage, ErrorCode.NATION_ADDRESS_INVALID);
+        if (!isSuperAdmin()&& !form.getUserId().equals(getCurrentUser())){
+            throw new BadRequestException("You don't have permission to create this address");
         }
-        // 2. Check User tồn tại
         User user = userRepository.findById(form.getUserId())
                 .orElseThrow(() -> new NotFoundException("User not found", ErrorCode.USER_ERROR_NOT_FOUND));
         Nation province = nationRepository.findById(form.getProvinceId())
@@ -71,9 +56,8 @@ public class AddressController extends ABasicController{
         Nation commune = nationRepository.findById(form.getCommuneId())
                 .orElseThrow(() -> new NotFoundException("Commune not found", ErrorCode.NATION_ERROR_NOT_FOUND));
 
-        // 4. Xử lý logic isDefault (Quan trọng)
+        //Set only one address at a time
         if (form.getIsDefault()) {
-            // Tìm và bỏ mặc định của tất cả địa chỉ cũ của User này
             addressRepository.unsetDefaultByUserId(user.getId());
         }
         Address address = addressMapper.fromCreateFormToEntity(form);
@@ -81,7 +65,6 @@ public class AddressController extends ABasicController{
         address.setProvince(province);
         address.setDistrict(district);
         address.setCommune(commune);
-
         addressRepository.save(address);
         apiMessageDto.setMessage("Create Nation success");
         return apiMessageDto;
@@ -90,19 +73,12 @@ public class AddressController extends ABasicController{
     @PutMapping(value = "/update", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('A-U')")
     public ApiMessageDto<String> update(@Valid @RequestBody UpdateAddressForm form, BindingResult bindingResult) {
-        if (bindingResult.hasErrors()) {
-            String errorMessage = bindingResult.getFieldErrors()
-                    .stream()
-                    .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                    .collect(Collectors.joining(", "));
-            throw new BadRequestException(errorMessage, ErrorCode.NATION_ADDRESS_INVALID);
+        if (!isSuperAdmin()&& !form.getUserId().equals(getCurrentUser())){
+            throw new BadRequestException("You don't have permission to create this product");
         }
-
-        // 1. Tìm address cũ
         Address address = addressRepository.findById(form.getId())
                 .orElseThrow(() -> new NotFoundException("Address not found", ErrorCode.ADDRESS_ERROR_NOT_FOUND));
-
-        // 2. Xử lý logic isDefault (Nếu set cái này là default thì gỡ các cái khác của User đó)
+        //Unset default with old address
         if (form.getIsDefault() != null && form.getIsDefault()) {
             addressRepository.unsetDefaultByUserId(address.getUser().getId());
         }
@@ -127,38 +103,53 @@ public class AddressController extends ABasicController{
         return apiMessageDto;
     }
 
-    // --- GET DETAIL ---
     @GetMapping(value = "/get/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('A-G')")
     public ApiMessageDto<AddressDto> get(@PathVariable("id") Long id) {
+        if(!isSuperAdmin()){
+            throw new BadRequestException("You don't have permission to get", ErrorCode.USER_ERROR_PERMISSION);
+        }
         Address address = addressRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Address not found", ErrorCode.ADDRESS_ERROR_NOT_FOUND));
-
         AddressDto addressDto = addressMapper.fromEntityToDto(address);
-
+        ApiMessageDto<AddressDto> apiMessageDto = new ApiMessageDto<>();
+        apiMessageDto.setData(addressDto);
+        apiMessageDto.setMessage("Get address success");
+        return apiMessageDto;
+    }
+    @GetMapping(value = "/getMyAddress", produces = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('A-G')")
+    public ApiMessageDto<AddressDto> getMyAddress() {
+        Address address = addressRepository.findById(getCurrentUser())
+                .orElseThrow(() -> new NotFoundException("Address not found", ErrorCode.ADDRESS_ERROR_NOT_FOUND));
+        AddressDto addressDto = addressMapper.fromEntityToDto(address);
         ApiMessageDto<AddressDto> apiMessageDto = new ApiMessageDto<>();
         apiMessageDto.setData(addressDto);
         apiMessageDto.setMessage("Get address success");
         return apiMessageDto;
     }
 
-    // --- DELETE ---
     @DeleteMapping(value = "/delete/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('A-D')")
     public ApiMessageDto<String> delete(@PathVariable("id") Long id) {
+        Long currentUserID = getCurrentUser();
         Address address = addressRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Address not found", ErrorCode.ADDRESS_ERROR_NOT_FOUND));
-
+        User userCurrent  = address.getUser();
+        if(getKind() !=1 && !currentUserID.equals(userCurrent.getId())){
+            throw new BadRequestException("You don't have permission to delete address", ErrorCode.ADDRESS_ERROR_PERMISSION);
+        }
         addressRepository.delete(address);
         ApiMessageDto<String> apiMessageDto = new ApiMessageDto<>();
         apiMessageDto.setMessage("Delete Address success");
         return apiMessageDto;
     }
-
-    // --- LIST (PHÂN TRANG & FILTER) ---
     @GetMapping(value = "/list", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('A-L')")
     public ApiMessageDto<ResponseListDto<List<AddressDto>>> list(AddressCriteria criteria, Pageable pageable) {
+        if(getKind() !=1){
+            throw new BadRequestException("You don't have permission to get list", ErrorCode.ADDRESS_ERROR_PERMISSION);
+        }
         // 1. Query từ DB
         Page<Address> page = addressRepository.findAll(criteria.getSpecification(), pageable);
